move tests to new directory
